+++ huginn-proxy-lib/src/proxy/handler/mod.rs
@@ -1,7 +1,9 @@
 pub mod header_manipulation;
 pub mod headers;
+pub mod host;
 pub mod rate_limit_validation;
 pub mod request;
 pub use headers::{add_forwarded_headers, akamai_header_value, tls_header_value};
+pub use host::{extract_request_host_inner, strip_host_port};
 pub use rate_limit_validation::check_rate_limit;
-pub use request::{extract_request_host_inner, handle_proxy_request, strip_host_port};
+pub use request::handle_proxy_request;
